Dropbox exceptions implemented

+++ dropbox_encryptor.py
@@ -1,4 +1,5 @@
 import shutil
+from dataSekura_exceptions import DropboxDownloadException, DropboxUploadException
 from db_module import Db_object
 from encryptor import Encryptor
 from local_encryptor import Local_encryptor
@@ -12,31 +13,39 @@ class DB_encryptor(Encryptor):
 
     def encrypt(self,folder,pwd,enc,hash,fs):
         cwd = os.getcwd()
-        folder_path, folder_metadata = self.db.download_folder_launch(folder)
+        try:
+            folder_path, folder_metadata = self.db.download_folder_launch(folder)
+        except Exception:
+            raise DropboxDownloadException()
         os.chdir(cwd)
-        self.folderDict = self.local.encrypt(folder_path,pwd,enc,hash,fs)
-        if self.folderDict == -1:
+        try:
+            self.folderDict = self.local.encrypt(folder_path,pwd,enc,hash,fs)
+        except Exception as e:
             shutil.rmtree(folder_path)
-            return
-        print("Cleaning residual files...")
-        self.db.upload_file(self.folderDict['volume_path'],folder_metadata.path_display+".bin")
-        self.fs.delete_vol(self.folderDict["volume_path"])
+            raise e
+        try:
+            self.db.upload_file(self.folderDict['volume_path'],folder_metadata.path_display+".bin")
+        except Exception:
+            raise DropboxUploadException()
+        if os.path.isfile(self.folderDict["volume_path"]):
+            self.fs.delete_vol(self.folderDict["volume_path"])
         self.db.remove_folder(folder)
-        print("Dropbox Folder Successfully Encrypted!")
     
     def decrypt(self,file,path,pwd):
         cwd = os.getcwd()
         full_path = os.path.abspath(file)
         file_noext = self.fs.remove_file_extension(full_path)
         os.chdir(cwd)
-        self.folderDict = self.local.decrypt(file_noext,pwd)
-        if self.folderDict == -1:
+        try:
+            self.folderDict = self.local.decrypt(file_noext,pwd)
+        except Exception as e:
             os.chdir(cwd)
             os.remove(file)
-            return
-        print("Restoring folder contents...")
-        self.db.upload_folder(self.fs.remove_file_extension(path),file)
-        print("Cleaning up residual files...")
+            raise e
+        try:    
+            self.db.upload_folder(self.fs.remove_file_extension(path),file)
+        except Exception:
+            raise DropboxUploadException()
         self.db.remove_bin(path)
-        shutil.rmtree(file_noext)
-        print("Dropbox Folder Successfully Decrypted!")
+        if os.path.isdir(file_noext):
+            shutil.rmtree(file_noext)
